+++ Libraries/InternalGearRatioValue.h
@@ -3,6 +3,6 @@
 
 NS_CORE_START
 
-USER_DEFINED_TYPE_FLOATING( InternalGearRatioValue, double )
+USER_DEFINED_TYPE_FLOATING( InternalGearRatioValue, float )
 
 NS_CORE_END
