+++ SintezPPDefK/DefK.cpp
@@ -20,17 +20,17 @@ void ari::DefK::calcExample()
 	NS_CORE Log::showValue( "Количество передач:", NS_CORE Singletons::getInstance()->getInitialData()._numberOfGears );
 
 	const auto &initialData = NS_CORE Singletons::getInstance()->getInitialData();
-	
+
 	const size_t testKSize = NS_CORE Singletons::getInstance()->getSettings()->getDefKSettings()._testsCount;
-	std::vector<NS_CORE InternalGearRatioValueArray> kValues(testKSize);
+	std::vector<NS_CORE InternalGearRatioValueArray> kValues( testKSize );
 	for ( size_t i = 0; i < initialData._numberOfPlanetaryGears; i++ )
 	{
-		for ( int j = 0; j < testKSize; j++ )
-			kValues[j].emplace_back( NS_CORE InternalGearRatioValue( ( rand() % 30 ) / 10.0 + 2 ) );
+		for ( size_t j = 0; j < testKSize; j++ )
+			kValues[j].emplace_back( NS_CORE InternalGearRatioValue( ( rand() % 30 ) / 10.0f + 2 ) );
 	}
 
 	std::vector<ari::InternalGearRatios> initialK;
-	for ( int j = 0; j < testKSize; j++ )
+	for ( size_t j = 0; j < testKSize; j++ )
 		initialK.emplace_back( NS_CORE InternalGearRatios( kValues[j] ) );
 
 	NS_CORE Code code;
@@ -46,12 +46,12 @@ void ari::DefK::calcExample()
 			if ( iSize > 0 )
 			{
 				int unique = 1;
-				for ( int i = 0; i < iSize - 1; i++ )
+				for ( size_t i = 0; i < iSize - 1; i++ )
 				{
 					if ( realI[i] != NS_CORE RatioValue( 0 ) )
 					{
 						bool finded = false;
-						for ( int j = i + 1; j < iSize; j++ )
+						for ( size_t j = i + 1; j < iSize; j++ )
 						{
 							if ( realI[i] == realI[j] )
 							{
@@ -68,22 +68,22 @@ void ari::DefK::calcExample()
 				}
 				if ( unique == initialData._numberOfGears )
 				{
-					if (!isWrited)
+					if ( !isWrited )
 					{
-						NS_CORE Singletons::getInstance()->getIOFileManager()->writeToFile(NS_CORE IOFileManager::eOutputFileType::K_TEST, code);
-						NS_CORE Singletons::getInstance()->getIOFileManager()->writeToFile(NS_CORE IOFileManager::eOutputFileType::DEF_K_LOG, code);
+						NS_CORE Singletons::getInstance()->getIOFileManager()->writeToFile( NS_CORE IOFileManager::eOutputFileType::K_TEST, code );
+						NS_CORE Singletons::getInstance()->getIOFileManager()->writeToFile( NS_CORE IOFileManager::eOutputFileType::DEF_K_LOG, code );
 						isWrited = true;
 					}
 					NS_CORE Singletons::getInstance()->getIOFileManager()->writeToFile( NS_CORE IOFileManager::eOutputFileType::DEF_K_LOG, realI );
 					NS_CORE Singletons::getInstance()->getIOFileManager()->writeToFile( NS_CORE IOFileManager::eOutputFileType::DEF_K_LOG, k );
 				}
 			}
-		}	
+		}
 	}
 }
 
 void DefK::run()
-{ 
+{
 	showParams();
 
 	if ( NS_CORE Singletons::getInstance()->getSettings()->getDefKSettings()._doTest )
@@ -100,13 +100,13 @@ void DefK::run()
 		NS_CORE InternalGearRatioArray	 ans = solveSimple.calculate( code );
 		NS_CORE Log::log( ".", false );
 
-		if ( !NS_CORE Singletons::getInstance()->getSettings()->getDefKSettings()._calcKQuick && ans.size() == 0)
+		if ( !NS_CORE Singletons::getInstance()->getSettings()->getDefKSettings()._calcKQuick && ans.size() == 0 )
 		{
 			NS_CORE Log::log( "#", false );
 			DefKSelection solveSelection;
 			ans = solveSelection.calculate( code );
 		}
-		
+
 
 		bool isWrited = false;
 		const size_t ansSize = ans.size();
@@ -127,11 +127,11 @@ void DefK::run()
 				//NS_CORE Singletons::getInstance()->getIOFileManager()->writeToFile( NS_CORE IOFileManager::eOutputFileType::DONE_K, realI );
 				NS_CORE Singletons::getInstance()->getIOFileManager()->writeToFile( NS_CORE IOFileManager::eOutputFileType::DEF_K_LOG, realI );
 				NS_CORE Singletons::getInstance()->getIOFileManager()->writeToFile( NS_CORE IOFileManager::eOutputFileType::DEF_K_LOG, ans[i] );
-					
+
 			}
 		}
-		
-	}	
+
+	}
 	system( "pause" );
 }
