+++ SintezPPKinCharacteristics/SolveFunctionDiv.h
@@ -10,13 +10,13 @@ class SolveFunctionDiv;
 
 typedef std::shared_ptr<SolveFunctionDiv> SolveFunctionDiv_p;
 
-class SolveFunctionDiv: public ISolveFunction
+class SolveFunctionDiv : public ISolveFunction
 {
 public:
 
 	static SolveFunctionDiv_p					create();
 
-	virtual double calc( const IFunction_p func, const double intervalBeg, const double intervalEnd ) override;
+	virtual float calc( const IFunction_p func, const float intervalBeg, const float intervalEnd ) override;
 };
 
 NS_ARI_END
